\name{MixedTS-package}
\alias{MixedTS-package}
\alias{MixedTS}
\docType{package}
\title{
Mixed Tempered Stable Distribution
}
\description{
This package provides detailed functions for univariate Mixed Tempered Stable distribution distribution with Gamma density. This distribution encompasses, Variance Gamma and Symmetric Geo-Stable as special cases. The package contains routine for mle estimation, for the computation of density, probability, quantile and random numbers.}
\details{
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Lorenzo Mercuri, Edit Rroji.
}
\references{
Rroji E. and Mercuri L.(2014)"Mixed Tempered Stable Distribution" ArXiv e-printsArXiv e-prints.
http://arxiv.org/abs/1405.7603 .

Rroji E.(2013):"Risk Attribution in semi-heavy tailed distributions" PhD Thesis. 
http://boa.unimib.it/bitstream/10281/49833/1/Phd_unimib_076099.pdf .
}

\keyword{ Mixed Tempered Stable, Tempered Stable, Geo-Stable, Variance Gamma }
