\name{MixedTS-package}
\alias{MixedTS-package}
\alias{MixedTS}
\docType{package}
\title{
Mixed Tempered Stable Distribution
}
\description{
This package provides detailed functions for univariate Mixed Tempered Stable distribution distribution with Gamma density. This distribution encompasses, Variance Gamma and Symmetric Geo-Stable as special cases. The package contains routine for mle estimation, for the computation of density, probability, quantile and random numbers
}
\details{
\tabular{ll}{
Package: \tab MixedTS\cr
Type: \tab Package\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Lorenzo Mercuri, Edit Rroji

Maintainer: Lorenzo Mercuri <lorenzo.mercuri@unimi.it>
}
\references{
 Barndorff-Nielsen,O.E., Kent,J. and Sorensen, M. (1982): Normal variance-mean mixtures and z-distributions, \emph{International Statistical Review}, 50, 145-159.

Kuchler, U. and Tappe, S. (2014): Exponential stockmodels driven by tempered stable processes. \emph{Journal of Econometrics},181 (1), 53-63.

Madan, D.B. and Seneta E. (1990): The variance gamma (V.G.) model for share market returns, \emph{Journal of Business}, 63, 511-524

Rroji, E and Mercuri, L.(2014): Mixed Tempered Stable distribution \emph{UNIMI-Research Papers in Economics, Business, and Statistics}, 64.
}
